+++ webapp/static/js/state.js
@@ -2,6 +2,7 @@ export const state = {
   projects: [],
   selectedProjectId: "",
   documents: [],
+  chatMessages: [],
   documentFilter: "",
   assessmentSession: null,
   assessmentQuestion: null,